adjust SortKey to have fragmentKey instead of gapKey
with FragmentIndex to compare fragment keys
